@@ -1,50 +1,21 @@
-/**
- * Copyright (c) 2017-present, Smartcar, Inc. All rights reserved.
-
- * You are hereby granted a limited, non-exclusive, worldwide, royalty-free
- * license to use, copy, modify, and distribute this software in source code or
- * binary form, for the limited purpose of this software's use in connection
- * with the web services and APIs provided by Smartcar.
- *
- * As with any software that integrates with the Smartcar platform, your use of
- * this software is subject to the Smartcar Developer Agreement. This copyright
- * notice shall be included in all copies or substantial portions of the software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
 package com.smartcar.sdk;
 
+import android.content.Context;
 import android.net.Uri;
+import android.view.View;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mockito;
 import org.robolectric.RobolectricTestRunner;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @RunWith(RobolectricTestRunner.class)
 public class SmartcarAuthTest {
-
-    @Test
-    public void smartcarAuth_generateUrl() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-
-        String requestUri = smartcarAuth.generateUrl();
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&approval_prompt=auto&mode=live";
-
-        assertEquals(expectedUri, requestUri);
-    }
 
     @Test
     public void smartcarAuth_authUrlBuilder() {
@@ -59,8 +30,25 @@ public class SmartcarAuthTest {
 
 
         SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String requestUri = smartcarAuth.new AuthUrlBuilder()
-                .build();
+        String requestUri = smartcarAuth.authUrlBuilder().build();
+
+        assertEquals(expectedUri, requestUri);
+    }
+
+    @Test
+    public void smartcarAuth_authUrlBuilder_testMode() {
+        String clientId = "client123";
+        String redirectUri = "scclient123://test";
+        String redirectUriEncoded = "scclient123%3A%2F%2Ftest";
+        String[] scope = {"read_odometer", "read_vin"};
+        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code" +
+                "&client_id=" + clientId +
+                "&redirect_uri=" + redirectUriEncoded +
+                "&mode=test&scope=read_odometer%20read_vin";
+
+
+        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, true, null);
+        String requestUri = smartcarAuth.authUrlBuilder().build();
 
         assertEquals(expectedUri, requestUri);
     }
@@ -74,15 +62,15 @@ public class SmartcarAuthTest {
         String vin = "1234567890ABCDEFG";
         String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code" +
                 "&client_id=" + clientId +
-                "&redirect_uri=" + redirectUriEncoded + 
+                "&redirect_uri=" + redirectUriEncoded +
                 "&mode=live&scope=read_odometer%20read_vin" +
-                "&approval_prompt=force&make=BMW&state=some%20state" + 
+                "&approval_prompt=force&make=BMW&state=some%20state" +
                 "&single_select=true" +
                 "&single_select_vin=" + vin;
 
         SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String requestUri = smartcarAuth.new AuthUrlBuilder()
-                .setApprovalPrompt(true)
+        String requestUri = smartcarAuth.authUrlBuilder()
+                .setForcePrompt(true)
                 .setMakeBypass("BMW")
                 .setState("some state")
                 .setSingleSelect(true)
@@ -93,123 +81,100 @@ public class SmartcarAuthTest {
     }
 
     @Test
-    public void smartcarAuth_generateUrl_forcePrompt() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
+    public void smartcarAuth_addClickHandler() {
 
-        String requestUri = smartcarAuth.generateUrl(true);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&approval_prompt=force&mode=live";
+        // Setup mocks
+        Context context = mock(Context.class);
+        View view = mock(View.class);
 
-        assertEquals(expectedUri, requestUri);
+        // Execute methods
+        SmartcarAuth smartcarAuth = new SmartcarAuth(
+                "client123",
+                "scclient123://test",
+                new String[] {"read_odometer", "read_vin"},
+                null
+        );
+        smartcarAuth.addClickHandler(context, view);
+
+        // Verify mocks
+        verify(view, times(1))
+                .setOnClickListener(Mockito.any(View.OnClickListener.class));
+
     }
 
     @Test
-    public void smartcarAuth_generateUrl_state() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
+    public void smartcarAuth_addClickHandler_withAuthUrl() {
 
-        String requestUri = smartcarAuth.generateUrl("somestring");
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&state=somestring&approval_prompt=auto&mode=live";
+        // Setup mocks
+        Context context = mock(Context.class);
+        View view = mock(View.class);
 
-        assertEquals(expectedUri, requestUri);
+        // Execute methods
+        SmartcarAuth smartcarAuth = new SmartcarAuth(
+                "client123",
+                "scclient123://test",
+                new String[] {"read_odometer", "read_vin"},
+                null
+        );
+        String authUrl = smartcarAuth.authUrlBuilder()
+                .setState("foo")
+                .setMakeBypass("TESLA")
+                .setSingleSelect(false)
+                .build();
+        smartcarAuth.addClickHandler(context, view, authUrl);
+
+        // Verify mocks
+        verify(view, times(1))
+                .setOnClickListener(Mockito.any(View.OnClickListener.class));
+
     }
 
     @Test
-    public void smartcarAuth_generateUrl_authVehicleInfo() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String make = "TESLA";
-        VehicleInfo authVehicleInfo = new VehicleInfo.Builder().make(make).build();
+    public void smartcarAuth_launchAuthFlow() {
 
-        String requestUri = smartcarAuth.generateUrl(authVehicleInfo);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&approval_prompt=auto&mode=live&make=" + make;
+        // Setup mocks
+        Context context = mock(Context.class);
 
-        assertEquals(expectedUri, requestUri);
+        // Execute methods
+        SmartcarAuth smartcarAuth = new SmartcarAuth(
+                "client123",
+                "scclient123://test",
+                new String[] {"read_odometer", "read_vin"},
+                null
+        );
+
+        smartcarAuth.launchAuthFlow(context);
+
     }
 
     @Test
-    public void smartcarAuth_generateUrl_stateAndForcePrompt() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
+    public void smartcarAuth_launchAuthFlow_withAuthUrl() {
 
-        String requestUri = smartcarAuth.generateUrl("somestring", true);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&state=somestring&approval_prompt=force&mode=live";
+        // Setup mocks
+        Context context = mock(Context.class);
 
-        assertEquals(expectedUri, requestUri);
-    }
+        // Execute methods
+        SmartcarAuth smartcarAuth = new SmartcarAuth(
+                "client123",
+                "scclient123://test",
+                new String[] {"read_odometer", "read_vin"},
+                null
+        );
+        String authUrl = smartcarAuth.authUrlBuilder()
+                .setState("foo")
+                .setMakeBypass("TESLA")
+                .setSingleSelect(false)
+                .build();
+        smartcarAuth.launchAuthFlow(context, authUrl);
 
-    @Test
-    public void smartcarAuth_generateUrl_stateAndAuthVehicleInfo() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String make = "TESLA";
-        VehicleInfo vehicleInfo = new VehicleInfo.Builder().make(make).build();
 
-        String requestUri = smartcarAuth.generateUrl("somestring", vehicleInfo);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&state=somestring&approval_prompt=auto&mode=live&make=" + make;
-
-        assertEquals(expectedUri, requestUri);
-    }
-
-    @Test
-    public void smartcarAuth_generateUrl_forcePromptAndAuthVehicleInfo() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String make = "TESLA";
-        VehicleInfo vehicleInfo = new VehicleInfo.Builder().make(make).build();
-
-        String requestUri = smartcarAuth.generateUrl(true, vehicleInfo);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&approval_prompt=force&mode=live&make=" + make;
-
-        assertEquals(expectedUri, requestUri);
-    }
-
-    @Test
-    public void smartcarAuth_generateUrl_stateAndforcePromptAndAuthVehicleInfo() {
-        String clientId = "client123";
-        String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
-        SmartcarAuth smartcarAuth = new SmartcarAuth(clientId, redirectUri, scope, null);
-        String make = "TESLA";
-        VehicleInfo vehicleInfo = new VehicleInfo.Builder().make(make).build();
-
-        String requestUri = smartcarAuth.generateUrl("somestring", true, vehicleInfo);
-        String expectedUri = "https://connect.smartcar.com/oauth/authorize?response_type=code&client_id="
-                + clientId + "&redirect_uri=" + redirectUri + "&scope=" + scope +
-                "&state=somestring&approval_prompt=force&mode=live&make=" + make;
-
-        assertEquals(expectedUri, requestUri);
     }
 
     @Test
     public void smartcarAuth_receiveResponse() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
@@ -225,7 +190,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_mismatchRedirectUri() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
         String wrongRedirectUri = "wrongscheme://test";
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
@@ -242,7 +207,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_nullUri() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
@@ -258,7 +223,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_nullCode() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
@@ -274,7 +239,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_accessDenied() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
             public void handleResponse(SmartcarResponse smartcarResponse) {
@@ -290,7 +255,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_vehicleIncompatible() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
             public void handleResponse(SmartcarResponse smartcarResponse) {
@@ -306,7 +271,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_vehicleIncompatibleWithVehicle() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
             public void handleResponse(SmartcarResponse smartcarResponse) {
@@ -329,7 +294,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_nullCodeWithMessage() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
@@ -345,7 +310,7 @@ public class SmartcarAuthTest {
     public void smartcarAuth_receiveResponse_codeWithState() {
         String clientId = "client123";
         String redirectUri = "scclient123://test";
-        String scope = "read_odometer read_vin";
+        String[] scope = {"read_odometer", "read_vin"};
 
         new SmartcarAuth(clientId, redirectUri, scope, new SmartcarCallback() {
             @Override
